requirements/local.txt
# Local development dependencies go here
-r base.txt

#coverage==4.4.2
#django-coverage-plugin==1.5.0

django-extensions==1.9.8
#Werkzeug==0.14.1
#django-test-plus==1.0.22
#factory-boy==2.9.2

django-debug-toolbar==1.9.1

# improved REPL
#ipdb==0.10.3


